# implementation/openapi.yaml
...
openapi: 3.1.0
info:
  title: CommerceLens 商脉经营分析 API（当前实现）
  version: '1.0'
  description: 对应 commerce profile 的实现契约。源设计保留在 docs/reference/commerce-lens/contracts/openapi.yaml。业务时区 Asia/Shanghai；日期左闭右开；金额为人民币分的十进制字符串，比率为十进制字符串。所有响应对象额外字段在本契约显式列出。运行事件
    schemaVersion=1.0。会话登录、权限、查询与运行均由真实后端执行；本地确定性规划器不代表已完成在线大模型评测。
servers:
- url: /api/commerce/v1
security:
- sessionCookie: []
paths:
  /me:
    get:
      operationId: getMe
      summary: 当前身份与授权范围
      tags:
      - me
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Me'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /overview:
    get:
      operationId: getOverview
      summary: 授权范围的经营总览
      tags:
      - overview
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Overview'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: storeIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 96
        style: form
        explode: true
        description: 可重复参数或逗号分隔店铺 ID；任何显式无权店铺拒绝整次请求。
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: endExclusive
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: comparison
        in: query
        required: true
        schema:
          enum:
          - PREVIOUS_WEEK_SAME_DAYS
          - PREVIOUS_PERIOD
...
  /queries:
    post:
      operationId: queryMetrics
      summary: 执行严格受控的指标查询
      tags:
      - queries
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/QueryResult'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
        description: 先 GET /auth/session 获取。登录成功后使用响应的新 csrfToken；所有 POST/PATCH/DELETE 均要求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                requestedStoreIds:
                  type: array
                  items:
                    type: string
                    minLength: 1
                    maxLength: 96
                  minItems: 0
                  description: 省略或空列表使用当前全部授权店铺；不得混入无权店铺。
                spec:
                  $ref: '#/components/schemas/QuerySpec'
              required:
              - spec
      security:
      - sessionCookie: []
        csrfToken: []
  /runs:
    post:
      operationId: createRun
      summary: 创建诊断运行，连接与执行独立
      tags:
      - runs
      responses:
        '202':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Run'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
        description: Same tenant/subject/operation/key with different body returns 409.
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
        description: 先 GET /auth/session 获取。登录成功后使用响应的新 csrfToken；所有 POST/PATCH/DELETE 均要求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRun'
      security:
      - sessionCookie: []
        csrfToken: []
    get:
      operationId: listRuns
      summary: 当前用户可见运行
      tags:
      - runs
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    type: object
                    additionalProperties: false
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Run'
                      nextCursor:
                        type:
                        - string
                        - 'null'
                    required:
                    - items
                    - nextCursor
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: status
        in: query
        required: false
        schema:
          enum:
          - QUEUED
          - PLANNING
          - WAITING_APPROVAL
          - RUNNING
          - CANCEL_REQUESTED
          - SUCCEEDED
          - PARTIAL
          - FAILED
          - CANCELLED
          - EXPIRED
          - ACCESS_REVOKED
  /runs/{runId}:
    get:
      operationId: getRun
      summary: 运行快照
      tags:
      - runs
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Run'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
  /runs/{runId}/approval:
    post:
      operationId: approveRun
      summary: 批准或拒绝指定版本计划
      tags:
      - runs
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Run'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
        description: 先 GET /auth/session 获取。登录成功后使用响应的新 csrfToken；所有 POST/PATCH/DELETE 均要求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Approval'
      security:
      - sessionCookie: []
        csrfToken: []
  /runs/{runId}/cancel:
    post:
      operationId: cancelRun
      summary: 幂等请求取消
      tags:
      - runs
      responses:
        '202':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Run'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
        description: 先 GET /auth/session 获取。登录成功后使用响应的新 csrfToken；所有 POST/PATCH/DELETE 均要求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                expectedRunVersion:
                  type: integer
                  minimum: 1
              required:
              - expectedRunVersion
      security:
      - sessionCookie: []
        csrfToken: []
  /runs/{runId}/events:
    get:
      operationId: subscribeRunEvents
      summary: SSE订阅与重放，不启动任务
      tags:
      - runs
      responses:
        '200':
          description: 'Ordered persistent events; heartbeat comments have no ID. Event payload follows RunEvent. Expired
            cursor: 410.'
          content:
            text/event-stream:
              schema:
                type: string
              example: 'id: 1

                event: run.created

                data: {"schemaVersion":"1.0","runId":"r1","sequence":1,"type":"run.created","occurredAt":"2026-09-21T01:00:00Z","payload":{}}


                '
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          type: string
          pattern: ^[0-9]+$
      - name: after
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: runId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
      description: 750ms 扫描持久事件，15s 注释心跳。Last-Event-ID 优先于 after；仅重放/订阅，不创建或推进运行。终态事件后关闭连接。
  /evidence/{evidenceId}:
    get:
      operationId: getEvidence
      summary: 读取当前仍有权访问的证据
      tags:
      - evidence
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Evidence'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: evidenceId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
  /reports:
    get:
      operationId: listReports
      summary: 当前可见报告
      tags:
      - reports
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    type: object
                    additionalProperties: false
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Report'
                      nextCursor:
                        type:
                        - string
                        - 'null'
                    required:
                    - items
                    - nextCursor
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
  /reports/{reportId}:
    get:
      operationId: getReport
      summary: 读取不可变报告
      tags:
      - reports
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Report'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
  /reports/{reportId}/export:
    get:
      operationId: exportReport
      summary: 重新鉴权后导出Markdown
      tags:
      - reports
      responses:
        '200':
          description: Download report with Content-Disposition attachment. Export is audited.
          content:
            text/markdown:
              schema:
                type: string
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
      description: 返回 Markdown 文本及 attachment 文件名 commerce-report.md，每次导出重新鉴权并登记审计。
...
  /anomaly-rules/{ruleId}:
    patch:
      operationId: updateRule
      summary: CAS更新规则并产生新版本
      tags:
      - anomaly-rules
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/AnomalyRule'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: ruleId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
        description: 先 GET /auth/session 获取。登录成功后使用响应的新 csrfToken；所有 POST/PATCH/DELETE 均要求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                expectedVersion:
                  type: integer
                  minimum: 1
                definition:
                  $ref: '#/components/schemas/AnomalyRuleInput'
              required:
              - expectedVersion
              - definition
      security:
      - sessionCookie: []
        csrfToken: []
...
  /ingestions/{jobId}:
    get:
      operationId: getIngestion
      summary: 查询导入校验进度
      tags:
      - ingestions
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  requestId:
                    type: string
                    minLength: 1
                    maxLength: 96
                  data:
                    $ref: '#/components/schemas/Ingestion'
                required:
                - requestId
                - data
        default:
          description: 'Error: 400/401/403/404/409/410/422/429/503 as documented'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 96
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Spring WebFlux SESSION Cookie；HttpOnly、SameSite=Strict、8 小时会话。开发 HTTP 默认 Secure=false；HTTPS 部署设 COMMERCE_COOKIE_SECURE=true。登录前
        GET /auth/session 先建立匿名会话及 CSRF token。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 必须匹配同一个 SESSION 的 CSRF token。
  schemas:
...
    Store:
      type: object
      additionalProperties: false
      properties:
        storeId:
          type: string
          minLength: 1
          maxLength: 96
        name:
          type: string
        platform:
          type: string
      required:
      - storeId
      - name
      - platform
    Me:
      type: object
      additionalProperties: false
      properties:
        subjectId:
          type: string
          minLength: 1
          maxLength: 96
        tenantId:
          type: string
          minLength: 1
          maxLength: 96
        displayName:
          type: string
        roles:
          type: array
          items:
            enum:
            - TENANT_ADMIN
            - OPS_MANAGER
            - STORE_OPERATOR
            - VIEWER
        authzVersion:
          type: integer
        stores:
          type: array
          items:
            $ref: '#/components/schemas/Store'
      required:
      - subjectId
      - tenantId
      - displayName
      - roles
      - authzVersion
      - stores
...